requirements.txt
django
psycopg2
gunicorn
djangorestframework
